feat: new prompt library and prompt composition
