+++ backend/src/gemini/gemini.service.ts
@@ -10,6 +10,9 @@ import { ApiLog, Lesson, ScenarioEvaluation } from '@/types';
 import { Timestamp } from 'firebase-admin/firestore';
 import { ApilogService } from '../apilog/apilog.service';
 import { performance } from "perf_hooks"; // For timing
+import { CLOZE_SYSTEM_PROMPT, buildClozeUserMessage } from '../prompts/cloze.prompts';
+import { buildKanjiDetailsSystemPrompt, buildKanjiDetailsUserMessage } from '../prompts/vocab.prompts';
+import { buildScenarioEvaluatorPrompt } from '../prompts/evaluation.prompts';
 
 @Injectable()
 export class GeminiService implements OnModuleInit {
@@ -561,8 +564,8 @@ export class GeminiService implements OnModuleInit {
   } // end generateQuestion
 
   async generateKanjiDetails(kanji: string) {
-    const systemPrompt = `You are a Japanese Kanji expert. Provide metadata for the kanji: ${kanji}.`;
-    const userMessage = `Generate a valid JSON object matching the Kanji Alive API structure for the character ${kanji}.`;
+    const systemPrompt = buildKanjiDetailsSystemPrompt(kanji);
+    const userMessage = buildKanjiDetailsUserMessage(kanji);
 
     const schema = {
       type: 'OBJECT',
@@ -688,53 +691,8 @@ export class GeminiService implements OnModuleInit {
     // Strip inline furigana (format: 漢字[ふりがな]) before sending to the AI and using in fallbacks.
     const cleanSentence = sentence.replace(/\[[^\]]*\]/g, '');
 
-    const systemPrompt = `You are an expert Japanese morphological analysis assistant specializing in cloze generation for SRS language learning flashcards.
-
-**Task:**
-Given a target vocabulary word in dictionary form and a context sentence that uses it (possibly conjugated or inflected), produce a cloze-deleted version of the sentence by replacing the occurrence of the target word with [_____].
-
-**Rules:**
-1. Identify the surface form: locate the token(s) in the sentence that represent the conjugated or inflected occurrence of the target vocabulary.
-2. Replace the ENTIRE conjugated word form — including its conjugation endings and any directly fused auxiliary verbs — with [_____]. This covers:
-   - All verb conjugations: plain/polite, past/non-past, negative, potential, causative, passive (食べる → 食べた, 食べます, 食べません, 食べられる, etc.)
-   - Te-form auxiliary constructions attached to the stem (食べている, 食べてみる, 食べてしまった → the entire compound is the target's occurrence)
-   - Suru-verb nominals used verbally (勉強する → 勉強しました → replace the whole 勉強しました)
-   - I-adjective conjugations (大きい → 大きかった, 大きくない → the whole conjugated form)
-   - Na-adjective citation form WITHOUT the connector な (有名な歌手 → replace only 有名, leaving な intact: [_____]な歌手)
-3. Preserve ALL other characters in the sentence EXACTLY. Do NOT rewrite, paraphrase, reorder, simplify, or otherwise alter any part of the sentence outside the blank.
-4. If the target word appears more than once, replace ONLY the first occurrence.
-5. If the target word is embedded in a morphologically related compound that has a distinct meaning (e.g., target: 食べる but sentence contains 大食い), do NOT replace it — these are different lexical items. Scan only for direct inflectional derivatives.
-6. If no occurrence can be confidently identified even after morphological analysis, replace the longest token in the sentence that shares the most morphemes with the target word. Never return the sentence unchanged.
-7. The placeholder MUST be written as exactly: [_____] — one opening bracket, five underscores, one closing bracket. No other format is acceptable.
-8. Do NOT add Romaji, parenthetical notes, translations, or any annotation to the output.
-9. Do NOT output any text before or after the JSON object. Do NOT use markdown code blocks or backticks.
-
-**Examples:**
-Vocabulary: "食べる"
-Sentence: "朝ごはんにパンを食べました。"
-Output: { "clozeSentence": "朝ごはんにパンを[_____]。" }
-
-Vocabulary: "行く"
-Sentence: "明日、学校に行きます。"
-Output: { "clozeSentence": "明日、学校に[_____]。" }
-
-Vocabulary: "食べる"
-Sentence: "毎日、野菜を食べています。"
-Output: { "clozeSentence": "毎日、野菜を[_____]。" }
-
-Vocabulary: "勉強する"
-Sentence: "毎日日本語を勉強しています。"
-Output: { "clozeSentence": "毎日日本語を[_____]。" }
-
-Vocabulary: "有名"
-Sentence: "彼は有名な歌手です。"
-Output: { "clozeSentence": "彼は[_____]な歌手です。" }
-
-Vocabulary: "大きい"
-Sentence: "その犬はとても大きかったです。"
-Output: { "clozeSentence": "その犬はとても[_____]。" }`;
-
-    const userMessage = `Vocabulary: "${targetVocab}"\nSentence: "${cleanSentence}"`;
+    const systemPrompt = CLOZE_SYSTEM_PROMPT;
+    const userMessage = buildClozeUserMessage(targetVocab, cleanSentence);
 
     const schema = {
       type: 'OBJECT',
@@ -936,54 +894,7 @@ Output: { "clozeSentence": "その犬はとても[_____]。" }`;
     const logRef = await this.apilogService.startLog(initialLogData);
 
     try {
-      // Improved Transcript Formatting
-
-      const historyText = chatHistory.map((m) => {
-        // Check against the internal ID 'user', not the display name
-        const isUser = m.speaker === 'user';
-
-        // Map internal ID to the specific Scenario Role
-        const currentRoleName = isUser ? scenarioContext.userRole : scenarioContext.aiRole;
-
-        // Label it clearly for the LLM
-        return `[${isUser ? 'User' : 'AI'} - ${currentRoleName}]: ${m.text}`;
-      }).join("\n");
-
-      const prompt = `
-        Act as a strict Japanese language teacher.
-        Evaluate this Roleplay Session based on:
-        1. Did they achieve the Goal? (success: true/false): (${scenarioContext.goal})
-        2. Grammar/Vocabulary Usage (Level ${scenarioContext.difficulty})
-        3. Rate naturalness (1-5).
-        4. Provide specific feedback in English.
-        5. Identify 1-3 critical mistakes.
-
-        SCENARIO CONTEXT:
-        - Title: ${scenarioContext.title}
-        - Goal: ${scenarioContext.goal}
-        - User Role: ${scenarioContext.userRole || 'User'} (The person being evaluated)
-        - AI Role: ${scenarioContext.aiRole || 'AI'} (The conversation partner)
-        - OBJECTIVE MET: ${scenarioContext.isObjectiveMet}
-
-        TRANSCRIPT:
-        ${historyText}
-
-        **CRITICAL GRADING RULES:**
-        1. **Output Language:** Write the 'feedback' and 'explanation' fields in **ENGLISH ONLY**. Do not use Japanese for the report text.
-        2. **Failure Condition:** If 'OBJECTIVE MET' is 'NO', the 'success' field MUST be false, and the 'rating' MUST be 1 or 2 (Fail). Do not give high ratings for incomplete work.
-        3. **Success Condition:** If 'OBJECTIVE MET' is 'YES', grade based on naturalness and grammar.
-        4. **Ignore Punctuation:** the user input is a speech transcript. Do NOT critique missing commas, periods, or quotation marks. Focus ONLY on word choice, grammar, and natural flow.
-
-        INSTRUCTION: Evaluate the USER (${scenarioContext.userRole || 'User'}) based on how well they achieved the goal. Do not confuse the User's responses with the AI's responses. 
-        OUTPUT LANGUAGE: the feedback and explanation must be in English. Do not write the report in Japanese but do not translate the user's or AI's responses.
-        
-        Provide a structured evaluation in JSON using the attached schema.
-      `;
-
-      // LOG PROMPT for debugging
-      console.log('--- SCENARIO EVALUATION PROMPT ---');
-      console.log(prompt);
-      console.log('----------------------------------');
+      const prompt = buildScenarioEvaluatorPrompt(chatHistory, scenarioContext);
 
       const response = await this.client.models.generateContent({
         model: this.modelName,
